+++ src/parsagon/__init__.py
@@ -1,1 +1,2 @@
-from parsagon.main import create, update, detail, run, batch_runs, delete, get_product, get_review_article, get_article_list
+from parsagon.main import create, update, detail, delete, get_product, get_review_article, get_article_list
+from parsagon.runs import run, batch_runs
