+++ src/parsagon/assistant.py
@@ -4,6 +4,7 @@ from parsagon.create import create_program
 from parsagon.executor import Executor
 from parsagon.print import assistant_print, assistant_spinner, browser_print, error_print
 from rich.prompt import Prompt
+from parsagon.runs import run, batch_runs
 
 
 def assist(task, headless, infer):
@@ -36,6 +37,15 @@ def assist(task, headless, infer):
                     outputs.append(output)
                 elif name == "create_program":
                     result = create_program(args["description"], headless=headless, infer=infer)
+                    output["output"] = json.dumps(result)
+                    outputs.append(output)
+                elif name == "run_program":
+                    result = run(**args)
+                    output["output"] = json.dumps(result)
+                    outputs.append(output)
+                elif name == "batch_runs":
+                    batch_name = input("Please enter a name for the batch run (for saving of intermediate results): ")
+                    result = batch_runs(batch_name, **args)
                     output["output"] = json.dumps(result)
                     outputs.append(output)
             with assistant_spinner():
